requirements_gpu.txt
faster-whisper==1.0.3
onnxruntime-gpu==1.23.2
openai-whisper==20250625
piper-tts==1.3.0
scipy==1.15.3
torch==2.9.0
torchaudio==2.9.0
